MAIN-T-117 Create endpoint to send Restore Password email (#92)
* Add password reset feature with token generation and email notification

 - Implemented a password reset functionality that checks user existence, generates a reset password token, and sends an email notification with the reset password token.
 - Created relevant test cases and updated user service and user facade.
 - Added endpoint to reset password and updated JwtAuthorizationFilter for password reset endpoint.
 - Included new database table for storing reset password tokens.

* Add logging and exception handling to reset password process

In the PasswordFacade class, a log statement and exception handling have been added to the reset password function. Now, when the password reset email is sent, it's wrapped in a try/catch block to handle any potential exceptions. Additionally, a debug-level log is generated whenever a rest password token is created.

* Refactor EmailService and remove unused variables

Refactored the EmailService class to extract redundant code into a separate 'sendEmail' function. This makes the code easier to read and maintain. Additionally, removed unused variables in 'restore-password.html' and 'PasswordSpec.kt'.