MAIN-T-124 It's possible to crate document with empty name and description
Make document name mandatory and add non-empty checks

This commit introduces a change to make the document name field mandatory in the database and in the DocumentEntity class. It also includes new not-empty check constraints for uid in user table, the token in reset_password_token table, and the name in the document table.

Add validation for DocumentRequest and improve log messages

The DocumentRequest fields have been modified to include validation annotations preventing empty or null values, enhancing data integrity. Changes have also been made to improve the precision of logged messages, including the addition of a post-save log in DefaultDocumentService.kt. Lastly, the code constructs a DocumentRequest using apply method for more readability.

Merge-request: MAIN-MR-72
Merged-by: Hanna Kurhuzenkava <hanna_kurhuzenkava@epam.com>