MAIN-T-99 Send registration email async
Add event-driven asynchronous email notifications

Added an asynchronous event-driven mechanism for sending email notifications after user registration. This was accomplished by publishing a `UserRegistrationEvent` directly after the user is created. The event is then consumed by a listener, `DokyEventListener`, which triggers the sending of the email. To accommodate this, the Awaitility library has also been added to the project to allow for testing of asynchronous processes. The `build.gradle` file was updated accordingly.