MAIN-T-130 Adjust Database namings
Rename database tables to plural form
The commit converts the naming of database tables from singular to plural form. It includes altering the tables for UserEntity, ScheduledTaskEntity, ResetPasswordTokenEntity, and DocumentEntity. An SQL migration script is added to perform the renaming and reindexing on the database side. Also, an additional unique constraint was added to reset_password_tokens on the user column for better data integrity.

Merge-request: MAIN-MR-78
Merged-by: Hanna Kurhuzenkava <hanna.kurhuzenkava@outlook.com>