MAIN-T-122 Run update index more frequently
Refactor event system and add document indexing events

This commit refactors the event system for better clarity, renames the existing classes and introduces new functionality. A new event listener `SolrEventListener` has been added to update the Solr index when a document is updated. The document service now emits an event when a document is updated or created.

Merge-request: MAIN-MR-85
Merged-by: Hanna Kurhuzenkava <hanna.kurhuzenkava@outlook.com>