DOKY-156 Spring Mapping for front
Update `Node.js` settings in `build.gradle`

This commit adds a `distBaseUrl` configuration in the Node definition in `build.gradle.kts`, which was previously undefined. Additionally, it includes a full specification for the `Node.js` repository in `settings.gradle.kts`. These changes aim to improve the handling of `Node.js` dependencies during builds.