DOKY-170 Error-free response for password restoration with non-existent email (#112)
DOKY-170 Error-free response for password restoration with non-existent email

Refactor password reset to handle non-existent users gracefully and avoiding the exception. The reason is to not expose information about emails that are used by registered users.

Update Swagger documentation as endpoint does not return 404 if user with provided email does not exist.