DOKY-185 Update the workflow to publish Javadocs (#158)
Upgrade Dokka to 2.0.0.

Updated Dokka to the latest major version 2.0.0 for improved functionality. Added forced version resolution for Jackson dependencies and introduced a new `gradle.properties` file for Gradle configurations.

Additionally, removed obsolete Azure and Spring-related dependencies and bundles to streamline the build configuration.

Changed the working directory from `server` to the project root (`.`) in the Javadoc publish workflow. This ensures the Gradle Wrapper runs correctly with the specified tasks for generating documentation.